Remove directory to App - Redirect to app to index.js - Should show index js to the index html file
