+++ src/index.js
@@ -1,14 +1,3 @@
-import createElementContainer from './elements/element_container';
-import Header from './components/header';
-import 'css-reset-and-normalize';
-import './styles/global.css';
+import App from './container/App';
 
-const bodyContainer = () => {
-  const bodyContainerElement = createElementContainer('div', [
-    'body-container',
-  ]);
-  const navBar = Header();
-  bodyContainerElement.append(navBar);
-  return bodyContainerElement;
-};
-document.getElementById('root').appendChild(bodyContainer());
+document.getElementById('root').appendChild(App());
